% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accupol.instruments}
\alias{accupol.instruments}
\title{Meta-data on Instruments for the dataset of the ACCUPOL research project.}
\format{
A data frame (tibble) containing the characteristics of policy Targets (Instrument, instrument.id, Sector and Degree of intervention.
}
\source{
Accupol research project (217239): https://cordis.europa.eu/project/id/788941
}
\usage{
data(accupol.instruments)
}
\description{
A tidy dataset of the characterestics of Instruments for the ACCUPOL research project.
}
\details{
When using the dataset, please cite:

Fernández-i-Marín, X., Knill, C., Steinbacher, C., & Steinebach, Y. (2024). Bureaucratic quality and the gap between implementation burden and administrative capacities. American Political Science Review, 118(3), 1240-1260. doi:10.1017/S0003055423001090
}
\examples{
data(accupol.instruments)
str(accupol.instruments)
}
\references{
Fernández-i-Marín, X., Knill, C., Steinbacher, C., & Steinebach, Y. (2024). Bureaucratic quality and the gap between implementation burden and administrative capacities. American Political Science Review, 118(3), 1240-1260. doi:10.1017/S0003055423001090
}
\seealso{
\code{\link{accupol.v1}} for the full dataset.
}
\keyword{datasets}
