% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accupol.v1}
\alias{accupol.v1}
\title{Policy portfolios for the dataset of the ACCUPOL research project.}
\format{
A data frame (tibble) containing the coverage of state intervention in a specific policy space (presence = 1 or absence = 0) for the ACCUPOL research project. It covers 23 countries in two sectors (Environmental, with 49 Targets and 11 Instruments; and Social, with Targets and Instruments) between 1976 and 2018.
}
\source{
Accupol research project (217239): https://cordis.europa.eu/project/id/788941
}
\usage{
data(accupol.v1)
}
\description{
A tidy dataset of a policy portfolios in social and environmental policies for the ACCUPOL research project.
}
\details{
When using the dataset, please cite:

Fernández-i-Marín, X., Knill, C., Steinbacher, C., & Steinebach, Y. (2024). Bureaucratic quality and the gap between implementation burden and administrative capacities. American Political Science Review, 118(3), 1240-1260. doi:10.1017/S0003055423001090
}
\examples{
data(accupol.v1)
str(accupol.v1)
}
\references{
Fernández-i-Marín, X., Knill, C., Steinbacher, C., & Steinebach, Y. (2024). Bureaucratic quality and the gap between implementation burden and administrative capacities. American Political Science Review, 118(3), 1240-1260. doi:10.1017/S0003055423001090
}
\seealso{
\code{\link{accupol.instruments}} for meta data about Instrument and Target metadata.
}
\keyword{datasets}
