% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{configurations}
\alias{configurations}
\title{Calculate a configuration of instruments}
\usage{
configurations(M)
}
\arguments{
\item{M}{Matrix with the policy portfolio, where the first dimension contains instruments and the second contains targets.}
}
\value{
Vector with the distribution of configurations
}
\description{
Internal function used to assess the number of configurations of
instruments. Configurations are specific combinations of instruments thay
may appear in different targets.
}
