% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{PolicyPortfolios-package}
\alias{PolicyPortfolios}
\alias{PolicyPortfolios-package}
\title{Tools for managing, measuring and visualizing policy portfolios}
\description{
Tools for simplifying the creation and management of data structures suitable
for dealing with policy portfolios, that is, two-dimensional spaces of
of policy instruments and policy targets. It allows to generate measures of
portfolio characteristics and facilitates their visualization.
}
\references{
\url{http://xavier-fim.net/packages/PolicyPortfolios/}.

Fernández-i-Marín, X., Knill, C. & Steinebach, Y. (2021). Studying Policy Design Quality in Comparative Perspective. _American Political Science Review_, online first.

Knill, C., Schulze, K. & Tosun, J. (2012). Regulatory Policy Outputs and Impacts. Exploring a Complex Relationship. _Regulation & Governance_, 5(4), 427-444. doi:10.1111/j.1748-5991.2012.01150.x.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://xavier-fim.net/packages/PolicyPortfolios/}
  \item \url{https://github.com/xfim/PolicyPortfolios}
  \item Report bugs at \url{https://github.com/xfim/PolicyPortfolios/issues}
}

}
\author{
\strong{Maintainer}: Xavier Fernández i Marín \email{xavier.fim@gmail.com} (\href{https://orcid.org/0000-0002-9522-8870}{ORCID})

}
\keyword{internal}
