% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{PolicyPortfolios}
\alias{PolicyPortfolios}
\alias{policyportfolios}
\title{Tools for managing, measuring and visualizing policy portfolios}
\description{
Tools for simplifying the creation and management of data structures suitable
for dealing with policy portfolios, that is, two-dimensional spaces of
of policy instruments and policy targets. It allows to generate measures of
portfolio characteristics and facilitates their visualization.
}
\references{
\url{http://xavier-fim.net/packages/PolicyPortfolios/}.

Fernández-i-Marín, X., Knill, C. & Steinebach, Y. (2021). Studying Policy Design Quality in Comparative Perspective. _American Political Science Review_, online first.

Knill, C., Schulze, K. & Tosun, J. (2012). Regulatory Policy Outputs and Impacts. Exploring a Complex Relationship. _Regulation & Governance_, 5(4), 427-444. doi:10.1111/j.1748-5991.2012.01150.x.
}
