% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_array.R
\name{pp_array}
\alias{pp_array}
\title{Convert a tidy dataset into a matrix of policy portfolios}
\usage{
pp_array(D, return_matrix = FALSE)
}
\arguments{
\item{D}{Data frame in a tidy format with the following columns: "Country", "Sector", "Year", "Instrument", "Target" and "covered". "covered" is a binary identificator of whether the portfolio space is covered by policy intervention (1) or not (0). The remaining columns identify the case. Notice that "Year" is a numeric value, while the remaining 4 case identifiers are factors.}

\item{return_matrix}{Logical value indicating whether the output must be in the form of an array with Country, Sector and Year dimensions present (although equal to one), or a simplified matrix of Instruments and Targets. Defaults to FALSE.}
}
\value{
An array when return_single is FALSE (the default), or a two-dimensional matrix with Instruments and Targets as first and second dimensions, respectively.
}
\description{
Calculate measures of interest of a policy portfolio and plot it.
}
\examples{
data(PolicyPortfolio)
# Returns an array
A <- pp_array(P.education)
dim(A)

# Returns a matrix
M <- pp_array(subset(P.education, Country == "Syldavia" & Year == 2022),
  return_matrix = TRUE)
dim(M)
}
