% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_measures.R
\name{pp_measures}
\alias{pp_measures}
\title{Measures of interest of a policy portfolio}
\usage{
pp_measures(D, id = NULL)
}
\arguments{
\item{D}{Data frame in a tidy format with the following columns: "Country", "Sector", "Year", "Instrument", "Target" and "covered". "covered" is a binary identificator of whether the portfolio space is covered by policy intervention (1) or not (0). The remaining columns identify the case. Notice that "Year" is a numeric value, while the remaining 4 case identifiers are factors.}

\item{id}{A list with up to two elements, namely "Country", and "Year" indicating the specific identification characteristics of the portfolio(s) that must be processed. Defaults to NULL to process all portfolios.}
}
\value{
A tidy dataset containing the portfolio identificators (Country, Sector and Year) plus the Measure identificators (Measure and Measure.label) and the value of the portfolio characteristic.
}
\description{
Calculate measures of interest of a policy portfolio.
}
\examples{
data(PolicyPortfolio)
m.education <- pp_measures(P.education)
m.education

# Calculate portfolio measures for a restricted set of portfolios defined by a list.
m.energy <- pp_measures(P.energy, id = list(Country = "Borduria", Year = 2022))
m.energy
}
