% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P.education}
\alias{P.education}
\title{Simulated policy portfolio with fake data for the education sector.}
\format{
A data frame (tibble) containing the coverage of state intervention in a specific policy space (presence = 1 or absence = 0).
}
\usage{
data(P.education)
}
\description{
A tidy dataset of a policy portfolio in an imaginary sector (education), containing 3 countries, 10 Instruments and 15 Targets, for 11 years.
}
\examples{
data(P.education)
str(P.education)
}
\keyword{datasets}
