% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P.energy}
\alias{P.energy}
\title{Simulated policy portfolio with fake data for the energy sector.}
\format{
A data frame (tibble) containing the coverage of state intervention in a specific policy space (presence = 1 or absence = 0).
}
\usage{
data(P.energy)
}
\description{
A tidy dataset of a policy portfolio in an imaginary sector (energy), containing 3 countries, 15 Instruments and 25 Targets, for 11 years.
}
\examples{
data(P.energy)
str(P.energy)
}
\keyword{datasets}
