% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{PolicyPortfolios}
\alias{PolicyPortfolios}
\alias{policyportfolios}
\alias{PolicyPortfolios-package}
\title{Tools for managing, measuring and visualizing policy portfolios}
\description{
PolicyPortfolios simplifies the creation of data structures suitable
for dealing with policy portfolios, that is, two-dimensional spaces of
of policy instruments and policy targets. It allows to generates measures of
their characteristics and facilitates its visualization.
}
\references{
\url{http://xavier-fim.net/packages/PolicyPortfolios}.
}
